Reformat folder structure of API. Also separated big methods into smaller more functional methods.
